Add support for UAT environment in AppConfiguration
Introduced a new conditional block in the `AppConfiguration`
class to handle the "UAT" (User Acceptance Testing) environment.
When the application is running in the "UAT" environment, the
`DatabaseExtensions.IntializeDatabaseAsync(app)` method is
called asynchronously to initialize the database. This aligns
the behavior of the "UAT" environment with the "Development"
environment in terms of database setup. No changes were made
to the existing logic for "Testing" or "Development" environments.